+++ mod-downloader/modoperations.py
@@ -30,13 +30,19 @@ def refresh_mods(mods: ModManager, path: str):
 def copy_to_target(manager: ModManager, pack_mods: list[Mod], path: str):
     for mod_file in os.listdir(path):
         result = manager.get_mod_by_filename(mod_file)
+        dep_mods = []
+
+        for mod in pack_mods:
+            for dep in mod.depend_on:
+                if dep not in dep_mods:
+                    dep_mods.append(dep)
+
         if result is not None:
             mod_path = os.path.join(path, mod_file)
             if result.filename not in installed_mods_list() or result not in pack_mods or result.state != "install":
-                print(f"Removing: {mod_path}")
-                os.remove(mod_path)
-        else:
-            print(f"Removing: {mod_path}")
+                if result not in dep_mods:
+                    print(f"Removing: {mod_path}")
+                    os.remove(mod_path)
 
     for pack_mod in pack_mods:
         pack_mod.copy2(path)
